+++ src/lib.rs
@@ -3,7 +3,6 @@
 
 pub mod cli;
 pub mod daemon;
-pub mod logrotate;
 pub mod path;
 pub mod protocol;
 pub mod spec;